+++ embeds/TogglEmbeds.py
@@ -5,7 +5,7 @@ from discord.ext import commands
 from discord import app_commands
 import inspect
 
-from Classes.TogglFunctions import TogglFunctions
+from classes.TogglFunctions import TogglFunctions
 from abstract.EmbedsAbstract import EmbedsAbstract
 
 from dotenv import dotenv_values
